Move single-layer utils out of the util/ catch-all
Relocate util files whose placement obscured their layer, so the directory
matches the library each compiles into:
- resource utilities -> src/resource/ (CritterFrmResolver, PathUtils,
  ResourcePaths), beside the rest of gecko_resource.
- Qt/app helpers -> src/ui/ (Settings, QtDialogs, FrmThumbnailGenerator,
  ExternalEditorLauncher, ResourceInitializer).

Cross-cutting helpers (Constants, Types, Exceptions, UndoStack, StringUtils) and
the core utils (Coordinates, ProHelper, SpatialIndex) stay in util/. Pure file
moves + include/CMake path updates; the link-time layer boundaries are already
enforced by the library split, so this is discoverability only.
