+++ src/Application.cpp
@@ -11,8 +11,8 @@
 #include "version.h"
 #include "resource/GameResources.h"
 #include "state/loader/MapLoader.h"
-#include "util/Settings.h"
-#include "util/QtDialogs.h"
+#include "ui/Settings.h"
+#include "ui/QtDialogs.h"
 #include "ui/core/MainWindow.h"
 #include "ui/core/EditorWidget.h"
 #include "ui/widgets/LoadingWidget.h"
