+++ src/resource/CritterFrmResolver.cpp
