+++ src/resource/CritterFrmResolver.h
