+++ src/resource/PathUtils.h
