+++ src/resource/ResourcePaths.h
