+++ src/ui/ExternalEditorLauncher.cpp
