+++ src/ui/ExternalEditorLauncher.h
