+++ src/ui/FrmThumbnailGenerator.cpp
