+++ src/ui/FrmThumbnailGenerator.h
