+++ src/ui/ResourceInitializer.h
