+++ src/ui/core/MainWindow.cpp
@@ -27,9 +27,9 @@
 #include "pattern/PatternSerializer.h"
 #include "format/map/Map.h"
 #include "util/Types.h"
-#include "util/Settings.h"
-#include "util/QtDialogs.h"
-#include "util/ExternalEditorLauncher.h"
+#include "ui/Settings.h"
+#include "ui/QtDialogs.h"
+#include "ui/ExternalEditorLauncher.h"
 #include "reader/lst/LstReader.h"
 #include "format/lst/Lst.h"
 #include "format/map/MapObject.h"
