+++ src/ui/panels/FileBrowserPanel.cpp
@@ -1,7 +1,7 @@
 #include "FileBrowserPanel.h"
 #include "resource/GameResources.h"
-#include "util/QtDialogs.h"
-#include "util/Settings.h"
+#include "ui/QtDialogs.h"
+#include "ui/Settings.h"
 #include "ui/dialogs/ProEditorDialog.h"
 #include "ui/theme/ThemeManager.h"
 #include "ui/UIConstants.h"
