+++ src/ui/rendering/RenderingEngine.cpp
@@ -7,7 +7,7 @@
 #include "resource/GameResources.h"
 #include "util/Constants.h"
 #include "util/ColorUtils.h"
-#include "util/ResourcePaths.h"
+#include "resource/ResourcePaths.h"
 #include "util/Coordinates.h"
 #include <spdlog/spdlog.h>
