+++ src/ui/widgets/GameLocationWidget.h
@@ -10,7 +10,7 @@
 #include <QGroupBox>
 #include <QProgressBar>
 #include <filesystem>
-#include "util/Settings.h"
+#include "ui/Settings.h"
 
 namespace geck {
